requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
sqlalchemy==2.0.45
pymysql==1.1.2
pydantic[email]==2.12.5
python-dotenv==1.1.1
